Setup tags and authors on CFGOVPage and convert them in events import
